sql
There is a mysql table containing cities of india with states . Write a function to

a. return a list of states in descending order of number of cities in the state

b. Returns a list of cities &amp; there states in alphabetical order of name of state

c. Returns a list of states with the number of cities in that state